refactor: restructure SubredditRow component for improved modularity
- Removed unused imports and types to streamline the component.
- Introduced new subcomponents: SubredditRowMain, SubredditRowActions, SubredditRowControls, and SubredditRowExpandedDetails for better separation of concerns.
- Updated the component to utilize the new structure while maintaining existing functionality.

This refactor enhances readability and maintainability of the SubredditRow component.
